add productRestController

+++ product/src/main/java/com/vnco/fusiontech/product/repository/AttributeRepository.java
@@ -4,6 +4,6 @@ import com.vnco.fusiontech.product.entity.Attribute;
 import com.vnco.fusiontech.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttributeRepository extends JpaRepository<Attribute, Long> {
+public interface AttributeRepository extends JpaRepository<Attribute, Integer> {
 
 }
