+++ product/src/main/java/com/vnco/fusiontech/product/repository/ProductAttributeRepository.java
@@ -4,6 +4,6 @@ import com.vnco.fusiontech.product.entity.Product;
 import com.vnco.fusiontech.product.entity.ProductAttribute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductAttributeRepository extends JpaRepository<ProductAttribute, Long> {
+public interface ProductAttributeRepository extends JpaRepository<ProductAttribute, Integer> {
 
 }
